Parser v1 restructure
