+++ src/utils.c
@@ -1,0 +1,24 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   utils.c                                            :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: alefranc <alefranc@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/08/25 17:48:18 by alefranc          #+#    #+#             */
+/*   Updated: 2022/08/25 17:52:04 by alefranc         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "cub3d.h"
+
+void	drain_fd(int fd)
+{
+	char *line;
+
+	while ((line = get_next_line(fd)) != NULL)
+	{
+		free(line);
+	}
+	close(fd);
+}
